Moved input sanitizer into a concern
